+++ src/main/java/mcjty/xnet/apiimpl/items/enums/ExtractMode.java
@@ -1,8 +1,8 @@
 package mcjty.xnet.apiimpl.items.enums;
 
 import mcjty.lib.varia.ComponentFactory;
+import mcjty.lib.gui.ITranslatableEnum;
 import mcjty.xnet.utils.I18nUtils;
-import mcjty.xnet.utils.ITranslatableEnum;
 
 public enum ExtractMode implements ITranslatableEnum<ExtractMode> {
     FIRST("xnet.enum.items.extractmode.first"),
