+++ src/main/java/mcjty/xnet/apiimpl/logic/enums/LogicMode.java
@@ -2,8 +2,8 @@ package mcjty.xnet.apiimpl.logic.enums;
 
 
 import mcjty.lib.varia.ComponentFactory;
+import mcjty.lib.gui.ITranslatableEnum;
 import mcjty.xnet.utils.I18nUtils;
-import mcjty.xnet.utils.ITranslatableEnum;
 
 
 public enum LogicMode implements ITranslatableEnum<LogicMode> {
