+++ src/main/java/mcjty/xnet/apiimpl/logic/enums/SensorMode.java
@@ -1,8 +1,8 @@
 package mcjty.xnet.apiimpl.logic.enums;
 
 import mcjty.lib.varia.ComponentFactory;
+import mcjty.lib.gui.ITranslatableEnum;
 import mcjty.xnet.utils.I18nUtils;
-import mcjty.xnet.utils.ITranslatableEnum;
 
 public enum SensorMode implements ITranslatableEnum<SensorMode> {
     OFF("xnet.enum.logic.sensormode.off"),
