--- src/main/java/mcjty/xnet/utils/ChoiceEvent.java
@@ -1,5 +1,0 @@
-package mcjty.xnet.utils;
-
-public interface ChoiceEvent<T> {
-    void choiceChanged(T newChoice);
-}
