--- src/main/java/mcjty/xnet/utils/ITranslatableEnum.java
@@ -1,11 +1,0 @@
-package mcjty.xnet.utils;
-
-public interface ITranslatableEnum<E extends Enum<E>> {
-
-    String name();
-    int ordinal();
-    String toString();
-
-    String getI18n();
-    String[] getI18nSplitedTooltip();
-}
